# copilot/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookYourCoach API",
        "description": "API REST pour la plateforme de réservation de cours avec coaches équestres ou autres sports. Cette API permet la gestion complète des utilisateurs (admin/enseignants/élèves), des réservations, des paiements et de la facturation.",
        "contact": {
            "name": "BookYourCoach Support",
            "url": "https://bookyourcoach.com",
            "email": "support@bookyourcoach.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur de développement BookYourCoach"
        }
    ],
    "paths": {
...
        "/api/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Liste des lieux",
                "description": "Récupère la liste complète des lieux de cours disponibles",
                "operationId": "getLocations",
                "responses": {
                    "200": {
                        "description": "Liste des lieux récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Créer un nouveau lieu",
                "description": "Crée un nouveau lieu de cours",
                "operationId": "createLocation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "city"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Centre Équestre de Paris"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de la Paix"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "75001"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 48.8566
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.3522
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Centre équestre moderne avec manèges couverts"
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "manège couvert",
                                            "carrière",
                                            "parking"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lieu créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Location"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lieu créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Détails d'un lieu",
                "description": "Récupère les détails d'un lieu spécifique",
                "operationId": "getLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du lieu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du lieu récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Location"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lieu non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Mettre à jour un lieu",
                "description": "Met à jour les informations d'un lieu",
                "operationId": "updateLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du lieu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Centre Équestre de Paris Rénové"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de la Paix"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "75001"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 48.8566
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.3522
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Centre équestre moderne avec nouveaux équipements"
                                    },
                                    "facilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "manège couvert",
                                            "carrière",
                                            "parking",
                                            "douches"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lieu mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Location"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lieu mis à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lieu non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Supprimer un lieu",
                "description": "Supprime un lieu",
                "operationId": "deleteLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du lieu",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lieu supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lieu supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lieu non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Location": {
                "title": "Location",
                "description": "Lieu de cours sur la plateforme",
                "properties": {
                    "id": {
                        "description": "Identifiant unique du lieu",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom du lieu",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Centre Équestre de Paris"
                    },
                    "address": {
                        "description": "Adresse complète",
                        "type": "string",
                        "maxLength": 500,
                        "example": "123 Rue de la Paix"
                    },
                    "city": {
                        "description": "Ville",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Paris"
                    },
                    "postal_code": {
                        "description": "Code postal",
                        "type": "string",
                        "maxLength": 20,
                        "example": "75001"
                    },
                    "country": {
                        "description": "Pays",
                        "type": "string",
                        "maxLength": 100,
                        "example": "France"
                    },
                    "latitude": {
                        "description": "Latitude GPS",
                        "type": "number",
                        "format": "float",
                        "example": 48.8566
                    },
                    "longitude": {
                        "description": "Longitude GPS",
                        "type": "number",
                        "format": "float",
                        "example": 2.3522
                    },
                    "description": {
                        "description": "Description du lieu",
                        "type": "string",
                        "example": "Centre équestre moderne"
                    },
                    "facilities": {
                        "description": "Équipements disponibles",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "manège couvert",
                            "carrière",
                            "parking"
                        ]
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "description": "Date de dernière mise à jour",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Bearer. Utilisez le token obtenu via /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}